# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "mechanic_service"
  description: "This is a API to manage a mechanic service"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint, (generates token)"
      description: "Validates customers credentials, generates and return a customer specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          
          examples:
            application/json:
              token:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NTEzOTIyMTksImlhdCI6MTc1MTM4ODYxOSwic3ViIjoiNyJ9.c8lPZCekbiTCSuUdtkbbIQeHM5D8A4Aq6aJ3obHf4rQ"
              message:"Loggin successful"
              status:"success"        

#==================create customer================================
  /customers: 
    post:
      tags:
       - Customers 
      summary: "Endpoint to create/register a new customer."
      description: "Endpoint to create a new customer ."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new Customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          descriptions: "Successfull Created  Customer"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
            
          examples:
            application/json:
              id: 1
              name: John Doe
              email: "johnd@email.com"
              address: "rocky mount"
              phone: "566 566 -6778"
              salary: 234555
              password: password123
 #==================get customers================================   
    get:
      tags:
       - Customers
      summary: "Returns all Customers"
      description: "Endpoint to retrieve a list of all customers."
      responses:
        200:
          description: "Retrieved Customers Successfully"           
          schema:
            $ref: "#/definitions/AllCustomers"   
#==================update customer================================
    put: 
      tags:
       - Customers 
      
      summary: "Endpoint to update a customer."
      description: "Endpoint to update a customer, this is a token authenticated route, and the customer must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a Customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          descriptions: "Successfully Updated customer"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              email: "johnd@email.com"
              address: "rocky mount"
              phone: "566 566 -6778"
              salary: 234555
              password: password123
    
#==================delete customer================================
  
    delete:
      tags:
       - Customers
      summary: "token to delete a customer"
      description: "token to delete a customer. Requires token authentication, customer must be logged in to delete their account"
      security:
        - bearerAuth: []
      schema:
            $ref: "#/definitions/DeleteCustomerResponse"
      responses:
        200:
          description: "Successfully Deleted customer"           
          examples:
            application/json:
              message: "customer id: 7, successfully deleted."
        404:
          description: "Customer not found"
#==================get specific customer================================
  /customers/{id}:
    get:
      tags:
       - Customers
      summary: Get customer by ID
      description: Returns details of a specific customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: int
          description: Unique ID of the customer
      responses:
        
        200:
          description: retrieve specific Customer 
          schema:
            $ref: '#/definitions/GetSpecificCustomer'
        
          examples:
                application/json:
                  id: 1
                  name: "John Doe"
                  email: "johnd@email.com"
                  address: "rocky mount"
                  phone: "566 566 -6778"
                  salary: 234555
                  password: "123"

   #==================mechanics==========================
   #==================create mechanic================================
...
  /inventory: 
    post:
      tags:
       - inventory
      summary: "Endpoint to create/register a new inventory."
      description: "Endpoint to create a new inventory ."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new inventory."
          required: true
          schema:
            $ref: "#/definitions/CreateInventoryPayload"
      responses:
        201:
          descriptions: "Successfull inventory Creation"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
            
          examples:
            application/json:
              id: 1
              name: "battery"
              price: 500.00
              
   #=========================get inventories============================== 
    get:
      tags:
       - inventory 
      summary: "Returns all inventory"
      description: "Endpoint to retrieve a list of all inventory."
      responses:
        200:
          description: "Retrieved inventory Successfully"           
          schema:
            $ref: "#/definitions/AllInventory" 

#=========================delete inventory==============================
  /inventory/{id}:
    delete:
      tags:
       - inventory 
      summary: Delete a inventory by ID
      operationId: deleteInventory  # Unique ID for the operation
      parameters:
        - in: path      # Parameter type (path parameter)
          name: id    # Parameter name
          required: true   # Parameter is required
          schema:
            type: integer  # Data type of the parameter
            format: int64   # Specific format for integer (e.g., 64-bit integer)
      responses:
        200:
          description: "Successfully Deleted inventory"           
          schema:
            $ref: "#/definitions/DeleteInventoryResponse"
          examples:
            application/json:
              message: "successfully deleted inventory <id>."
    
#=========================get specific inventory==============================
    get:
      tags:
       - inventory
      summary: Get inventory by ID
      description: Returns details of a specific inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: int
          description: Unique ID of the inventory
      responses:
        
        200:
          description: retrieve specific inventory 
          schema:
            $ref: '#/definitions/GetSpecificInventory'
        
          examples:
                application/json:
                  id: 1
                  name: "battery"
                  price: 500.00

    #=========================update inventory==============================  
    put:
      tags:
       - inventory
      summary: Update an inventory  by ID
      description: updateInventory
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the inventory item to update
          schema:
            type: integer
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/UpdateInventoryPayload'

      responses:
        200:
          description: Inventory item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/definitions/UpdateInventoryResponse'
          examples:
            application/json:
              id: 1
              name: "battery"
              price: 500.0
#=======================service_inventory=======================
#=======================ADD  inventory to service======================
...
definitions: 
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
      
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "int"
      password:
        type: "string"
      
    required:
      - name
      - email
      - address
      - phone
      - salary
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "int"
      name:
        type: "string"
      email:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "int"
      password:
        type: "string"
  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "int"
        name:
          type: "string"
        email:
          type: "string"
        address:
          type: "string"
        phone:
          type: "string"
        salary:
          type: "int"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "int"
      password:
        type: "string"
      
    required:
      - name
      - email
      - address
      - phone
      - salary
      - password
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "int"
      name:
        type: "string"
      email:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "int"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"   
  GetSpecificCustomer:
    items:
      type: object
      properties:
        id:
          type: "int"
        name:
          type: "string"
        email:
          type: "string"
        address:
          type: "string"
        phone:
          type: "string"
        salary:
          type: "int"

      required:
        - id
        - name
        - email
        - address
        - phone
        - salary
        - password
          

#==================mechanic================================
...
  CreateInventoryPayload:
    type: "object"
    properties:
      name:
        type: "string"
      price:
        type: "float"
      
      
    required:
      - name
      - price
  CreateInventoryResponse:
    type: "object"
    properties:
      id:
        type: "int"
      name:
        type: "string"
      price:
        type: "float"
...
  DeleteInventoryResponse:
    type: "object"
    properties:
      message:
        type: "string" 
  GetSpecificInventory:
    items:
      type: object
      properties:
        id:
          type: "int"
        name:
          type: "string"
        price:
          type: "float"
       
      required:
        - name
        - price
  UpdateInventoryPayload:
    type: "object"
    properties:
      name:
        type: "string"
      price:
        type: "number"
        format: "float"
        

    required:
      - name
      - price
  UpdateInventoryResponse:
    type: "object"
    properties:
      id:
        type: "int"
      name:
        type: "string"
      price:
        type: "number"
        format: "float"
        

  
 #===================mechanic_service======================
...